==> 5 - Classification models.ipynb <==
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score
from sklearn.metrics import f1_score
from sklearn.metrics import classification_report
from sklearn.tree import DecisionTreeClassifier
from sklearn import svm

from sklearn.model_selection import GridSearchCV
from keras.wrappers.scikit_learn import KerasClassifier

from keras.models import Sequential
from keras.layers import Dense
from sklearn.utils import shuffle
import pydotplus
import pandas as pd
import numpy as np

### Load original train and test set

DDoS_train = pd.read_csv('Train_test_data/DDoS_train.csv', sep=',') 
DDoS_test = pd.read_csv('Train_test_data/DDoS_test.csv', sep=',') 

PortScan_train = pd.read_csv('Train_test_data/PortScan_train.csv', sep=',') 
PortScan_test = pd.read_csv('Train_test_data/PortScan_test.csv', sep=',') 

Bot_train = pd.read_csv('Train_test_data/Bot_train.csv', sep=',')
Bot_test = pd.read_csv('Train_test_data/Bot_test.csv', sep=',') 

WebAttack_train = pd.read_csv('Train_test_data/WebAttack_train.csv', sep=',') 
WebAttack_test = pd.read_csv('Train_test_data/WebAttack_test.csv', sep=',') 

Brute_Force_train = pd.read_csv('Train_test_data/Brute_Force_train.csv', sep=',') 
Brute_Force_test = pd.read_csv('Train_test_data/Brute_Force_test.csv', sep=',') 

DoS_train = pd.read_csv('Train_test_data/Dos_train.csv', sep=',')
DoS_test = pd.read_csv('Train_test_data/Dos_test.csv', sep=',') 

### Load GAN balanced training set

GAN_Bot_train = pd.read_csv('GAN_generation\GAN_Balanced_Data\GAN_Bot_train.csv', sep=',')
GAN_WebAttack_train = pd.read_csv('GAN_generation\GAN_Balanced_Data\GAN_WebAttack_train.csv', sep=',')
GAN_Brute_Force_train = pd.read_csv('GAN_generation\GAN_Balanced_Data\GAN_Brute_Force_train.csv', sep=',')
GAN_DoS_train = pd.read_csv('GAN_generation\GAN_Balanced_Data\GAN_DoS_train.csv', sep=',')

WGAN_GP_Bot_train = pd.read_csv('GAN_generation\WGAN_GP_Balanced_Data\WGAN_GP_Bot_train.csv', sep=',')
WGAN_GP_WebAttack_train = pd.read_csv('GAN_generation\WGAN_GP_Balanced_Data\WGAN_GP_WebAttack_train.csv', sep=',')
WGAN_GP_Brute_Force_train = pd.read_csv('GAN_generation\WGAN_GP_Balanced_Data\WGAN_GP_Brute_Force_train.csv', sep=',')
WGAN_GP_DoS_train = pd.read_csv('GAN_generation\WGAN_GP_Balanced_Data\WGAN_GP_DoS_train.csv', sep=',')

### Load SMOTE balanced training set

SMOTE_Botnet_train = pd.read_csv('SMOTE_Balanced_Data\Botnet_smote_train.csv', sep=',')
SMOTE_WebAttack_train = pd.read_csv('SMOTE_Balanced_Data\WebAttack_smote_train.csv', sep=',')
SMOTE_Brute_Force_train = pd.read_csv('SMOTE_Balanced_Data\Brute_Force_smote_train.csv', sep=',')
SMOTE_DoS_train = pd.read_csv('SMOTE_Balanced_Data\Dos_smote_train.csv', sep=',')

### Original dataset X-Y train-test split

X_DDoS_train = DDoS_train.values[:,:-1]
X_DDoS_test = DDoS_test.values[:,:-1]

Y_DDoS_train = DDoS_train.values[:,-1:]
Y_DDoS_test = DDoS_test.values[:,-1:]


X_PortScan_train = PortScan_train.values[:,:-1]
X_PortScan_test = PortScan_test.values[:,:-1]

Y_PortScan_train = PortScan_train.values[:,-1:]
Y_PortScan_test = PortScan_test.values[:,-1:]


X_Bot_train = Bot_train.values[:,:-1]
X_Bot_test = Bot_test.values[:,:-1]

Y_Bot_train = Bot_train.values[:,-1:]
Y_Bot_test = Bot_test.values[:,-1:]


#Multi-class label

X_WebAttack_train = WebAttack_train.values[:,:-1]
X_WebAttack_test = WebAttack_test.values[:,:-1]

Y_WebAttack_train = WebAttack_train.values[:,-1:]
Y_WebAttack_test = WebAttack_test.values[:,-1:]


X_Brute_Force_train = Brute_Force_train.values[:,:-1]
X_Brute_Force_test = Brute_Force_test.values[:,:-1]

Y_Brute_Force_train = Brute_Force_train.values[:,-1:]
Y_Brute_Force_test = Brute_Force_test.values[:,-1:]


X_DoS_train = Dos_train.values[:,:-1]
X_DoS_test = Dos_test.values[:,:-1]

Y_DoS_train = Dos_train.values[:,-1:]
Y_DoS_test = Dos_test.values[:,-1:]

### Generated data X-Y train-test split

#GAN and WGAN-GP
X_GAN_Bot_train = GAN_Bot_train.values[:,:-1]
Y_GAN_Bot_train = GAN_Bot_train.values[:,-1:]

X_WGAN_GP_Bot_train = WGAN_GP_Bot_train.values[:,:-1]
Y_WGAN_GP_Bot_train = WGAN_GP_Bot_train.values[:,-1:]

X_GAN_WebAttack_train = GAN_WebAttack_train.values[:,:-1]
Y_GAN_WebAttack_train = GAN_WebAttack_train.values[:,-1:]

X_WGAN_GP_WebAttack_train = WGAN_GP_WebAttack_train.values[:,:-1]
Y_WGAN_GP_WebAttack_train = WGAN_GP_WebAttack_train.values[:,-1:]

X_GAN_Brute_Force_train = GAN_Brute_Force_train.values[:,:-1]
Y_GAN_Brute_Force_train = GAN_Brute_Force_train.values[:,-1:]

X_WGAN_GP_Brute_Force_train = WGAN_GP_Brute_Force_train.values[:,:-1]
Y_WGAN_GP_Brute_Force_train = WGAN_GP_Brute_Force_train.values[:,-1:]

X_GAN_DoS_train = GAN_DoS_train.values[:,:-1]
Y_GAN_DoS_train = GAN_DoS_train.values[:,-1:]

X_WGAN_GP_DoS_train = WGAN_GP_DoS_train.values[:,:-1]
Y_WGAN_GP_DoS_train = WGAN_GP_DoS_train.values[:,-1:]

#SMOTE
X_SMOTE_Bot_train = SMOTE_Botnet_train.values[:,:-1]
Y_SMOTE_Bot_train = SMOTE_Botnet_train.values[:,-1:]

X_SMOTE_WebAttack_train = SMOTE_WebAttack_train.values[:,:-1]
Y_SMOTE_WebAttack_train = SMOTE_WebAttack_train.values[:,-1:]

X_SMOTE_Brute_Force_train = SMOTE_Brute_Force_train.values[:,:-1]
Y_SMOTE_Brute_Force_train = SMOTE_Brute_Force_train.values[:,-1:]

X_SMOTE_DoS_train = SMOTE_DoS_train.values[:,:-1]
Y_SMOTE_DoS_train = SMOTE_DoS_train.values[:,-1:]

### Select the dataset for the models 

DATASET = DDoS_train

X_train = X_DDoS_train
X_test = X_DDoS_test

Y_train = Y_DDoS_train
Y_test = Y_DDoS_test

### ANN model for binary classification

model = Sequential()
model.add(Dense(128, input_dim=len(DATASET.columns)-1, activation='softmax'))
model.add(Dense(32, activation='relu'))
model.add(Dense(16, activation='relu'))
model.add(Dense(1, activation="sigmoid")) #sigmoid
# compile the keras model
model.compile(loss='binary_crossentropy', optimizer='Nadam', metrics=['accuracy']) #binary_crossentropy# fit the keras model on the dataset
model.fit(X_train, Y_train, epochs=30, batch_size=1000) #bot 20 epochs

### ANN Prediction for binary classification 

# predict probabilities for test set
yhat_probs = model.predict(X_test, verbose=1) 
# predict crisp classes for test set
yhat_classes = model.predict_classes(X_test, verbose=0) 
# reduce to 1d array
yhat_probs = yhat_probs[:, 0]
yhat_classes = yhat_classes[:, 0]

print(classification_report(Y_test, yhat_classes, digits=5))

### ANN model for multi-class classification

model = Sequential()
model.add(Dense(128, input_dim=len(DATASET.columns)-1, activation='softmax'))
model.add(Dense(32, activation='relu'))
model.add(Dense(16, activation='relu'))
model.add(Dense(3, activation="softmax")) # WebAttacks = 3 BruteForce = 3 Dos = 5
# compile the keras model
model.compile(loss='sparse_categorical_crossentropy', optimizer='Nadam', metrics=['accuracy'])
model.fit(X_train, Y_train, epochs=30, batch_size=1000)

### ANN Prediction for multi-class classification

y_pred = model.predict(X_test, verbose=1)
#y_pred
y_pred_bool = np.argmax(y_pred.tolist(), axis=1)
#y_pred_bool
print(classification_report(Y_test, y_pred_bool, digits=5))

### DT model

tree = DecisionTreeClassifier()
clf=tree.fit(X_train, Y_train)

### DT Prediction

# predict probabilities for test set
yhat_probs = clf.predict(X_test)
print(classification_report(Y_test, yhat_probs, digits=5))
